move shared migration msgs to migration package
