+++ contracts/snip721-migratable/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod contract;
 mod contract_migrate;
 pub mod msg;
+mod msg_untagged_deserializer;
 pub mod state;
 
 mod test_instantiate;
