+++ packages/migration/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod msg_types;
 pub mod state;
 pub mod execute;
+pub mod msg;
 
 mod test_execute;
